Move common composites to device_composites.py
